feat: simplify configuration for SPA applications
The initial implementation for SPAs required some unnecessary files to be created when running
production build: This pull request removes the requirement for a build:server step and the
existance of an ssr manifest when spa mode is set to `true`. This also adds a react-vanilla-spa
example to show users how to configure this mode.