fix(vue): preload links (#227)
Links are generated with `crossorigin` while the preloads aren't, giving a warning. This adds the `crossorigin` attribute to the preloads as well.

The resource http://localhost/assets/index-DtkymAZZ.css was preloaded using link preload but not used within a few seconds from the window's load event. Please make sure it has an appropriate `as` value and it is preloaded intentionally.