Add .vscode/settings.json (#236)
Since this repo uses a style that is not the default in either prettier nor biome, VSCode format on save results in huge modifications. This is very annoying and slows down development by a lot.

This settings file fixes that by using VSCode's built in formatter with the settings specified in `contributing.md`.